+++ system.actionStart.option2.lua
@@ -1,0 +1,8 @@
+-- Darkwinde START: system.actionStart(option 2)
+-- >> Option 2 is by default ALT+2
+
+
+-- Deactivate vertical stabilization on user request
+verticalEngines = not verticalEngines
+
+-- Darkwinde END: system.actionStart(option 2)